Refactor: Move KeywordContentEditCard under modules folder
