+++ src/modules/IdCardCreation/Step/KeywordContentStep.client.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { useFormContext } from 'react-hook-form';
 
-import { KeywordContentEditCard } from '~/modules/IdCardCreation/Card';
+import { KeywordContentEditCard } from '~/modules/KeywordContentEditCard';
 import { FormKeywordModel } from '~/types/idCard';
 
 const title = '나를 소개하는 키워드의\n 설명을 적어주세요!';
